Move convertIiifUriToInfoUri into the catalogue app
