+++ common/test/utils/convert-image-uri.test.ts
@@ -1,7 +1,4 @@
-import {
-  convertIiifUriToInfoUri,
-  convertImageUri,
-} from '../../utils/convert-image-uri';
+import { convertImageUri } from '../../utils/convert-image-uri';
 
 describe('convertImageUri for Prismic images', () => {
   it('creates a full-sized image', () => {
@@ -104,26 +101,6 @@ describe('convertImageUri for IIIF images', () => {
     );
     expect(result).toEqual(
       'https://iiif.wellcomecollection.org/image/b0005/full/300%2C/0/default.jpg'
-    );
-  });
-});
-
-describe('convertIiifUriToInfoUri', () => {
-  it('finds the info.json for a IIIF URI', () => {
-    const result = convertIiifUriToInfoUri(
-      'https://iiif.wellcomecollection.org/image/b0006.jpg/full/300%2C/0/default.jpg'
-    );
-    expect(result).toEqual(
-      'https://iiif.wellcomecollection.org/image/b0006.jpg/info.json'
-    );
-  });
-
-  it('finds the info.json for an unrecognised URI', () => {
-    const result = convertIiifUriToInfoUri(
-      'https://iiif.wellcomecollection.org/image/b0007.jp2'
-    );
-    expect(result).toEqual(
-      'https://iiif.wellcomecollection.org/image/b0007.jp2/info.json'
     );
   });
 });
